Created repository interfaces, and changed UserRepository to use Long for ID

+++ src/main/java/com/example/homeeconomics/economic/user/repository/UserRepository.java
@@ -3,4 +3,4 @@ package com.example.homeeconomics.economic.user.repository;
 import com.example.homeeconomics.economic.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer> { }
+public interface UserRepository extends JpaRepository<User, Long> {}
